move ComponentHarness DOM tests to run in JSDOM
